Refactor API key authentication to GigAPIKey class and integrate with NinjaAPI
